+++ packages/effect/dtslint/Either.ts
@@ -171,3 +171,33 @@ export type R = Either.Either.Right<typeof string$number>
 
 // $ExpectType string
 export type L = Either.Either.Left<typeof string$number>
+
+// -------------------------------------------------------------------------------------
+// do notation
+// -------------------------------------------------------------------------------------
+
+// $ExpectType Either<{ a: number; b: string; c: boolean; }, never>
+pipe(
+  Either.Do,
+  Either.bind("a", (
+    _scope // $ExpectType {}
+  ) => Either.right(1)),
+  Either.bind("b", (
+    _scope // $ExpectType { a: number; }
+  ) => Either.right("b")),
+  Either.let("c", (
+    _scope // $ExpectType { a: number; b: string; }
+  ) => true)
+)
+
+// $ExpectType Either<{ a: number; b: string; c: boolean; }, never>
+pipe(
+  Either.right(1),
+  Either.bindTo("a"),
+  Either.bind("b", (
+    _scope // $ExpectType { a: number; }
+  ) => Either.right("b")),
+  Either.let("c", (
+    _scope // $ExpectType { a: number; b: string; }
+  ) => true)
+)