+++ packages/effect/src/Option.ts
@@ -8,6 +8,7 @@ import type { LazyArg } from "./Function.js"
 import { constNull, constUndefined, dual, identity, isFunction } from "./Function.js"
 import type { TypeLambda } from "./HKT.js"
 import type { Inspectable } from "./Inspectable.js"
+import * as doNotation from "./internal/doNotation.js"
 import * as either from "./internal/either.js"
 import * as option from "./internal/option.js"
 import type { Order } from "./Order.js"
@@ -28,7 +29,7 @@ export type Option<A> = None<A> | Some<A>
  * @category symbols
  * @since 2.0.0
  */
-export const TypeId = Symbol.for("effect/Option")
+export const TypeId: unique symbol = Symbol.for("effect/Option")
 
 /**
  * @category symbols
@@ -1198,16 +1199,39 @@ export const exists: {
 // -------------------------------------------------------------------------------------
 
 /**
+ * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+ *
+ * Here's how the do simulation works:
+ *
+ * 1. Start the do simulation using the `Do` value
+ * 2. Within the do simulation scope, you can use the `bind` function to define variables and bind them to `Option` values
+ * 3. You can accumulate multiple `bind` statements to define multiple variables within the scope
+ * 4. Inside the do simulation scope, you can also use the `let` function to define variables and bind them to simple values
+ * 5. Regular `Option` functions like `map` and `filter` can still be used within the do simulation. These functions will receive the accumulated variables as arguments within the scope
+ *
+ * @see {@link Do}
+ * @see {@link bind}
+ * @see {@link let_ let}
+ *
+ * @example
+ * import { Option, pipe } from "effect"
+ *
+ * const result = pipe(
+ *   Option.Do,
+ *   Option.bind("x", () => Option.some(2)),
+ *   Option.bind("y", () => Option.some(3)),
+ *   Option.let("sum", ({ x, y }) => x + y),
+ *   Option.filter(({ x, y }) => x * y > 5)
+ * )
+ * assert.deepStrictEqual(result, Option.some({ x: 2, y: 3, sum: 5 }))
+ *
  * @category do notation
  * @since 2.0.0
  */
 export const bindTo: {
   <N extends string>(name: N): <A>(self: Option<A>) => Option<{ [K in N]: A }>
   <A, N extends string>(self: Option<A>, name: N): Option<{ [K in N]: A }>
-} = dual(
-  2,
-  <A, N extends string>(self: Option<A>, name: N): Option<{ [K in N]: A }> => map(self, (a) => ({ [name]: a } as any))
-)
+} = doNotation.bindTo<OptionTypeLambda>(map)
 
 const let_: {
   <N extends string, A extends object, B>(
@@ -1219,15 +1243,36 @@ const let_: {
     name: Exclude<N, keyof A>,
     f: (a: A) => B
   ): Option<{ [K in N | keyof A]: K extends keyof A ? A[K] : B }>
-} = dual(3, <A extends object, N extends string, B>(
-  self: Option<A>,
-  name: Exclude<N, keyof A>,
-  f: (a: A) => B
-): Option<{ [K in N | keyof A]: K extends keyof A ? A[K] : B }> =>
-  map(self, (a) => Object.assign({}, a, { [name]: f(a) }) as any))
+} = doNotation.let_<OptionTypeLambda>(map)
 
 export {
   /**
+   * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+   *
+   * Here's how the do simulation works:
+   *
+   * 1. Start the do simulation using the `Do` value
+   * 2. Within the do simulation scope, you can use the `bind` function to define variables and bind them to `Option` values
+   * 3. You can accumulate multiple `bind` statements to define multiple variables within the scope
+   * 4. Inside the do simulation scope, you can also use the `let` function to define variables and bind them to simple values
+   * 5. Regular `Option` functions like `map` and `filter` can still be used within the do simulation. These functions will receive the accumulated variables as arguments within the scope
+   *
+   * @see {@link Do}
+   * @see {@link bind}
+   * @see {@link bindTo}
+   *
+   * @example
+   * import { Option, pipe } from "effect"
+   *
+   * const result = pipe(
+   *   Option.Do,
+   *   Option.bind("x", () => Option.some(2)),
+   *   Option.bind("y", () => Option.some(3)),
+   *   Option.let("sum", ({ x, y }) => x + y),
+   *   Option.filter(({ x, y }) => x * y > 5)
+   * )
+   * assert.deepStrictEqual(result, Option.some({ x: 2, y: 3, sum: 5 }))
+   *
    * @category do notation
    * @since 2.0.0
    */
@@ -1235,6 +1280,32 @@ export {
 }
 
 /**
+ * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+ *
+ * Here's how the do simulation works:
+ *
+ * 1. Start the do simulation using the `Do` value
+ * 2. Within the do simulation scope, you can use the `bind` function to define variables and bind them to `Option` values
+ * 3. You can accumulate multiple `bind` statements to define multiple variables within the scope
+ * 4. Inside the do simulation scope, you can also use the `let` function to define variables and bind them to simple values
+ * 5. Regular `Option` functions like `map` and `filter` can still be used within the do simulation. These functions will receive the accumulated variables as arguments within the scope
+ *
+ * @see {@link Do}
+ * @see {@link bindTo}
+ * @see {@link let_ let}
+ *
+ * @example
+ * import { Option, pipe } from "effect"
+ *
+ * const result = pipe(
+ *   Option.Do,
+ *   Option.bind("x", () => Option.some(2)),
+ *   Option.bind("y", () => Option.some(3)),
+ *   Option.let("sum", ({ x, y }) => x + y),
+ *   Option.filter(({ x, y }) => x * y > 5)
+ * )
+ * assert.deepStrictEqual(result, Option.some({ x: 2, y: 3, sum: 5 }))
+ *
  * @category do notation
  * @since 2.0.0
  */
@@ -1248,14 +1319,35 @@ export const bind: {
     name: Exclude<N, keyof A>,
     f: (a: A) => Option<B>
   ): Option<{ [K in N | keyof A]: K extends keyof A ? A[K] : B }>
-} = dual(3, <A, N extends string, B>(
-  self: Option<A>,
-  name: Exclude<N, keyof A>,
-  f: (a: A) => Option<B>
-): Option<{ [K in keyof A | N]: K extends keyof A ? A[K] : B }> =>
-  flatMap(self, (a) => map(f(a), (b) => Object.assign({}, a, { [name]: b }) as any)))
+} = doNotation.bind<OptionTypeLambda>(map, flatMap)
 
 /**
+ * The "do simulation" in allows you to write code in a more declarative style, similar to the "do notation" in other programming languages. It provides a way to define variables and perform operations on them using functions like `bind` and `let`.
+ *
+ * Here's how the do simulation works:
+ *
+ * 1. Start the do simulation using the `Do` value
+ * 2. Within the do simulation scope, you can use the `bind` function to define variables and bind them to `Option` values
+ * 3. You can accumulate multiple `bind` statements to define multiple variables within the scope
+ * 4. Inside the do simulation scope, you can also use the `let` function to define variables and bind them to simple values
+ * 5. Regular `Option` functions like `map` and `filter` can still be used within the do simulation. These functions will receive the accumulated variables as arguments within the scope
+ *
+ * @see {@link bindTo}
+ * @see {@link bind}
+ * @see {@link let_ let}
+ *
+ * @example
+ * import { Option, pipe } from "effect"
+ *
+ * const result = pipe(
+ *   Option.Do,
+ *   Option.bind("x", () => Option.some(2)),
+ *   Option.bind("y", () => Option.some(3)),
+ *   Option.let("sum", ({ x, y }) => x + y),
+ *   Option.filter(({ x, y }) => x * y > 5)
+ * )
+ * assert.deepStrictEqual(result, Option.some({ x: 2, y: 3, sum: 5 }))
+ *
  * @category do notation
  * @since 2.0.0
  */