+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @effect/cli
 
+## 0.36.16
+
+### Patch Changes
+
+- Updated dependencies [[`e5e56d1`](https://github.com/Effect-TS/effect/commit/e5e56d138dbed3204636f605229c6685f89659fc)]:
+  - effect@3.1.1
+  - @effect/platform@0.52.1
+  - @effect/printer@0.33.10
+  - @effect/printer-ansi@0.33.10
+  - @effect/schema@0.66.13
+
+## 0.36.15
+
+### Patch Changes
+
+- Updated dependencies [[`9deab0a`](https://github.com/Effect-TS/effect/commit/9deab0aec9e99501f9441843e34df9afa10c5be9), [`7719b8a`](https://github.com/Effect-TS/effect/commit/7719b8a7350c14e952ffe685bfd5308773b3e271)]:
+  - @effect/platform@0.52.0
+
 ## 0.36.14
 
 ### Patch Changes